refactor: remove typing module and relocate PoolType protocol
- Delete [src/pyextremes/typing.py] to simplify package structure.
- Move [PoolType] protocol to [src/pyextremes/models/model_emcee.py] where it is primarily utilized.
- Update internal imports in [src/pyextremes/models/models.py] and [src/pyextremes/eva.py].
- Ensure PEP 8/20 compliance by co-locating typing definitions with their primary consumers.
