--- src/pyextremes/typing.py
@@ -1,11 +1,0 @@
-import typing
-
-
-@typing.runtime_checkable
-class PoolType(typing.Protocol):
-    def map(
-        self,
-        func: typing.Callable[..., typing.Any],
-        iterable: typing.Iterable[typing.Any],
-        /,
-    ) -> typing.Iterable[typing.Any]: ...
